requirements.txt
numpy
pandas
yfinance
arch
ssm-learn
tensorflow
tensorflow-probability
scikit-learn
xgboost
plotly
colorama